+++ src/modules/cmdpal/src/WindowsCommandPalette/Views/NoOpAction.xaml.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.ObjectModel;
+using System.Collections.Specialized;
+using System.ComponentModel;
+using System.Runtime.InteropServices;
+using DeveloperCommandPalette;
+using Microsoft.UI.Dispatching;
+using Microsoft.UI.Input;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Controls.Primitives;
+using Microsoft.UI.Xaml.Data;
+using Microsoft.UI.Xaml.Input;
+using Microsoft.UI.Xaml.Navigation;
+using Microsoft.Windows.CommandPalette.Extensions;
+using Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+namespace WindowsCommandPalette.Views;
+
+public sealed class NoOpAction : InvokableCommand
+{
+    public override ICommandResult Invoke()
+    {
+        return ActionResult.KeepOpen();
+    }
+}
